refactor: 장바구니 도메인에서 상품 서비스 호출해서 사용하도록 변경 (#134)
- 기존에 인터페이스를 호출하던 구조를 팀 컨벤션에 맞춰 서비스를 호출하도록 변경
