+++ src/main/java/com/example/i_commerce/domain/cart/service/CartQueryService.java
@@ -6,16 +6,17 @@ import com.example.i_commerce.domain.cart.controller.response.CartResponse;
 import com.example.i_commerce.domain.cart.controller.response.CartStoreGroupResponse;
 import com.example.i_commerce.domain.cart.entity.Cart;
 import com.example.i_commerce.domain.cart.entity.CartItem;
-import com.example.i_commerce.domain.cart.infrastructure.ProductClient;
-import com.example.i_commerce.domain.cart.infrastructure.ProductItemInfo;
+import com.example.i_commerce.domain.product.application.dto.ProductItemInfo;
 import com.example.i_commerce.domain.cart.repository.CartItemRepository;
 import com.example.i_commerce.domain.cart.repository.CartRepository;
+import com.example.i_commerce.domain.product.application.service.ProductQueryService;
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
+import java.util.function.Function;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -28,7 +29,7 @@ public class CartQueryService {
 
     private final CartRepository cartRepository;
     private final CartItemRepository cartItemRepository;
-    private final ProductClient productClient;
+    private final ProductQueryService productQueryService;
 
     public CartResponse getCartItems(Long userId) {
 
@@ -49,7 +50,11 @@ public class CartQueryService {
             .collect(Collectors.toSet());
 
         Map<Long, ProductItemInfo> productItemInfoMap =
-            productClient.getProductItemInfos(productItemIds);
+            productQueryService.getProductItemInfosByIds(productItemIds).stream()
+                .collect(Collectors.toMap(
+                    ProductItemInfo::productItemId,
+                    Function.identity()
+                ));
 
         Map<Long, StoreAccumulator> storeMap = new LinkedHashMap<>();
